+++ clang/lib/CodeGen/CodeGenFunction.h
@@ -3468,6 +3468,10 @@ public:
     /// True if lifetime markers should be used.
     bool UseLifetimeMarkers;
 
+    /// True if the variable was initialized in the entry block because its
+    /// declaration is bypassed by a goto or switch jump.
+    bool IsInitializedInEntryBlock;
+
     /// Address with original alloca instruction. Invalid if the variable was
     /// emitted as a global constant.
     RawAddress AllocaAddr;
@@ -3475,12 +3479,13 @@ public:
     struct Invalid {};
     AutoVarEmission(Invalid)
         : Variable(nullptr), Addr(Address::invalid()),
-          AllocaAddr(RawAddress::invalid()) {}
+          IsInitializedInEntryBlock(false), AllocaAddr(RawAddress::invalid()) {}
 
     AutoVarEmission(const VarDecl &variable)
         : Variable(&variable), Addr(Address::invalid()), NRVOFlag(nullptr),
           IsEscapingByRef(false), IsConstantAggregate(false),
-          UseLifetimeMarkers(false), AllocaAddr(RawAddress::invalid()) {}
+          UseLifetimeMarkers(false), IsInitializedInEntryBlock(false),
+          AllocaAddr(RawAddress::invalid()) {}
 
     bool wasEmittedAsGlobal() const { return !Addr.isValid(); }
 
@@ -3488,6 +3493,9 @@ public:
     static AutoVarEmission invalid() { return AutoVarEmission(Invalid()); }
 
     bool useLifetimeMarkers() const { return UseLifetimeMarkers; }
+    bool wasInitializedInEntryBlock() const {
+      return IsInitializedInEntryBlock;
+    }
 
     /// Returns the raw, allocated address, which is not necessarily
     /// the address of the object itself. It is casted to default
@@ -5486,6 +5494,8 @@ private:
 
   void emitZeroOrPatternForAutoVarInit(QualType type, const VarDecl &D,
                                        Address Loc);
+  LangOptions::TrivialAutoVarInitKind
+  getAutoVarInitKind(QualType type, const VarDecl &D);
 
 public:
   enum class EvaluationOrder {